Database work finished; Major bug fixes needed
All the database work is done however  a few major bugs became apparent when testing the code. 1. The friends page does not work as the friends collection believes that _friend is a Map and not a List<Friend> 2. Registration does not work as there is no way to make an account tied to an email. i.e. I cannot seem to fit the RegisterInfo page after email verification or on the registration page